EUD 를 제대로 — Memory 조건·동작, 오프셋 표, euddraft 빌드
화면과 CLI 양쪽에서 EUD 를 다룰 수 있게 했다.

셈을 바로잡았다. Deaths 표는 유닛 바깥·플레이어 안쪽이다 — 한 유닛이
12플레이어 x 4바이트를 차지하고 그런 항목이 228개다(0x0058A364, size 48,
length 228). 그전 코드는 `4 x (플레이어 x 228 + 유닛)` 으로 셈해 주소가
통째로 어긋나 있었다. EPD 나눗셈도 부호 있는 쪽이어야 한다 — 미네랄
0x0057F0F0 은 EPD -11421 이고, 부호 없이 나누면 1073730403 이 나온다.

CHK 안에서 EUD 는 그냥 Deaths 다. 플레이어 자리가 28 을 넘으면 EUD 로
보고(MappingCore 의 텍스트 트리거와 같은 잣대) 인자 자리표를 가상 종류
Memory / Memory Masked 것으로 바꿔 읽는다. 그러지 않으면 첫 자리가
"플레이어" 로 보이고, 거기 든 큰 수를 사람이 플레이어 번호로 고치는
순간 맵이 망가진다. 비트마스크 자리도 이제 보인다.

- 트리거 편집기: 종류 목록에 Memory / Memory Masked, 주소로 바로 편집,
  32비트를 다 쓰는 자리는 글자 상자로(QSpinBox 는 2^31 까지만 담는다)
- 주소 계산기: 주소 ↔ EPD ↔ Deaths 자리, 이름으로 찾기, 리마스터 지원 여부
- 트리거 › 이 맵의 EUD 훑기
- 트리거 › EUD 빌드 (euddraft) — epScript 편집까지
- splash-cli eud addr/epd/offsets/list/check/set-condition/set-action/build/which
- splash-cli trigger add/remove/duplicate/owners (없으면 EUD 를 넣을 자리가 없다)

오프셋 표는 붙박이 24개를 들고 있고, EUD Book 의 api.json 을 가리키면
935개까지 늘어난다. 그 파일은 라이선스 표기가 없어 저장소에 넣지 않았다.

저장 경로에서 하나 고쳤다. euddraft 가 뱉는 맵은 scenario.chk 를 로케일로
숨긴다 — 중립 로케일에 0바이트짜리 미끼를 두고 영어(1033)에 진짜를 넣는다.
MpqFile::getFile 은 열기에 실패했을 때만 다른 로케일을 보는데 빈 파일은
열기에 성공하므로 미끼가 잡혔다. readScenarioChk 가 중립 쪽이 비면 로케일을
훑는다. 이것이 없으면 splash-cli chk 가 0바이트를 내고 round-trip 검증이
빈 바이트끼리 견준다. 컬렉션의 한국 유즈맵 400개에는 이 꼴이 0개라 지금까지의
통과율에는 영향이 없다.

실측: 빈 트리거 맵에 EUD 조건·동작을 넣고 텍스트로 뽑았다 되돌려도 CHK
바이트가 같다. euddraft 로 구운 맵(트리거 0 -> 71, EUD 39군데)을 다시 열어
읽는다. 단위 테스트 362개, 화면 스모크 59개 통과.

+++ src/ui/trigger_editor.cpp
@@ -143,7 +143,7 @@ TriggerEditor::TriggerEditor(chk::MapDocument & document, io::GameGraphics & gra
     const auto triggerIndex = [this] { return static_cast<std::size_t>(currentIndex()); };
 
     connect(conditionArgs_, &TriggerArgumentPanel::typeChanged, this,
-            [this, triggerIndex](std::size_t slot, std::uint8_t type) {
+            [this, triggerIndex](std::size_t slot, std::uint32_t type) {
         if (currentIndex() < 0) return;
         const int row = conditions_->currentRow();
         if (document_.setConditionType(triggerIndex(), slot, type))
@@ -205,7 +205,7 @@ TriggerEditor::TriggerEditor(chk::MapDocument & document, io::GameGraphics & gra
     });
 
     connect(actionArgs_, &TriggerArgumentPanel::typeChanged, this,
-            [this, triggerIndex](std::size_t slot, std::uint8_t type) {
+            [this, triggerIndex](std::size_t slot, std::uint32_t type) {
         if (currentIndex() < 0) return;
         const int row = actions_->currentRow();
         if (document_.setActionType(triggerIndex(), slot, type))
